Drop DEPOT_PATH assumption
However, this replaces the assumption that the depot path is the same at
the call-site as on the scale-set vm with the assumption that the scale-set
vm has its depot path at the default location: /home/$user/.julia.  In addition,
this keeps the assumption that the vm and the caller use the same julia
version (major and minor).