Dynamic version
This is based off of the "elastic.jl" cluster manager
in the ClusterManagers.jl package, with some changes,
and adding in the Azure specific bits.  This means makes
the manager more dynamic so that machines are added to the
cluster as they are provisioined, and work can start on the
cluster as soon as the first machine is added to the cluster.